Rename minplus pkg to algebra + cleanups
