+++ java/de/uni_kl/cs/discodnc/curves/mpartc/pwaffine/MaxServiceCurve_MPARTC_PwAffine.java
@@ -29,6 +29,7 @@ package de.uni_kl.cs.discodnc.curves.mpartc.pwaffine;
 
 import ch.ethz.rtc.kernel.Curve;
 import ch.ethz.rtc.kernel.SegmentList;
+
 import de.uni_kl.cs.discodnc.curves.MaxServiceCurve;
 
 public class MaxServiceCurve_MPARTC_PwAffine extends Curve_MPARTC_PwAffine implements MaxServiceCurve {
